Moving ui components into ui package
