+++ vector/src/main/java/im/vector/app/features/roomprofile/polls/ended/RoomEndedPollsFragment.kt
@@ -19,7 +19,7 @@ package im.vector.app.features.roomprofile.polls.ended
 import dagger.hilt.android.AndroidEntryPoint
 import im.vector.app.R
 import im.vector.app.features.roomprofile.polls.RoomPollsType
-import im.vector.app.features.roomprofile.polls.list.RoomPollsListFragment
+import im.vector.app.features.roomprofile.polls.list.ui.RoomPollsListFragment
 
 @AndroidEntryPoint
 class RoomEndedPollsFragment : RoomPollsListFragment() {
